+++ IsekaiMod/Content/Features/IsekaiProtagonist/SecretPowerSelection.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 using static IsekaiMod.Main;
 
 namespace IsekaiMod.Content.Features.IsekaiProtagonist {
-    static class SecretPower {
+    static class SecretPowerSelection {
         private static readonly Sprite Icon_SecretPower = AssetLoader.LoadInternal(IsekaiContext, "Features", "ICON_SECRET_POWER.png");
         public static void Add() {
-            var SecretPower = Helpers.CreateBlueprint<BlueprintFeatureSelection>(IsekaiContext, "SecretPower", bp => {
+            var SecretPowerSelection = Helpers.CreateBlueprint<BlueprintFeatureSelection>(IsekaiContext, "SecretPowerSelection", bp => {
                 bp.SetName(IsekaiContext, "Secret Power");
                 bp.SetDescription(IsekaiContext, "On the verge of defeat, you were somehow able to draw out your secret power."
                     + "\nYou gain an additional Special Power.");